fix(app): Fix Application layer build errors
Fix 111 build errors across Application and Infrastructure layers:

## Method Signature Fixes
- Add GetByPeriodAsync(string period) overload to ICalculationRepository
  for callers passing period string like "2025-01"
- Add GetPendingAsync() to IApprovalRepository for admin view
- Add DeleteAsync() to IRoleRepository for role deletion
- Add LogActivityAsync() and GetAuditLogsAsync() to IAuditService

## Entity Property Fixes
- Fix Calculation.Achievement → Calculation.AchievementPercentage
- Add Department.Parent navigation property for hierarchy

## Type Conversion Fixes
- Cast double to decimal in CreateBand comparison (ReportGenerationService)
- Parse AuditAction from string in SecurityAuditService
- Convert AuditAction enum to string for SecurityEventDto.EventType
- Fix UserId from string to Guid? in AuditLogSearchQuery

## Missing Using Directives
- Add Domain.Enums to DashboardService, ReportGenerationService
- Add Domain.Enums and Audit.DTOs to SecurityAuditService

+++ src/Dorise.Incentive.Infrastructure/Reports/ReportGenerationService.cs
@@ -1,6 +1,7 @@
 using Dorise.Incentive.Application.Reports.DTOs;
 using Dorise.Incentive.Application.Reports.Services;
 using Dorise.Incentive.Domain.Entities;
+using Dorise.Incentive.Domain.Enums;
 using Dorise.Incentive.Domain.Interfaces;
 using Microsoft.Extensions.Logging;
 using System.Text;
@@ -703,7 +704,7 @@ public class ReportGenerationService : IReportGenerationService
 
     private static AchievementBandDto CreateBand(string name, double min, double max, IReadOnlyList<Calculation> calculations)
     {
-        var count = calculations.Count(c => c.AchievementPercentage.Value >= min && c.AchievementPercentage.Value < max);
+        var count = calculations.Count(c => c.AchievementPercentage.Value >= (decimal)min && c.AchievementPercentage.Value < (decimal)max);
         return new AchievementBandDto
         {
             BandName = name,
